+++ resources/properties.js
@@ -1,7 +1,7 @@
 'use strict';
 
 export default {
-    'version': '0.4.1-dev',
+    'version': '0.5.0-dev',
     'author_email': 'prometheussatyen@gmail.com',
 
     'eve_client_id': '537173a2901b47389695c27fbfa6f0b6',
@@ -83,4 +83,15 @@ export default {
         'ship': 300,
         'fatigue': 300,
     },
+
+    'contract_completed_statuses': [
+        'finished_issuer',
+        'finished_contractor',
+        'finished',
+        'cancelled',
+        'rejected',
+        'failed',
+        'deleted',
+        'reversed'
+    ],
 };