+++ src/components/views/Contracts.jsx
@@ -2,13 +2,35 @@
 
 import React from 'react';
 
+import {Card, CardHeader, CardText} from 'material-ui/Card';
+
 import ContractsTable from '../tables/ContractsTable';
 
-export default class SpFarming extends React.Component {
+const styles = {
+    card: {
+        margin: 10
+    }
+};
+
+export default class Contracts extends React.Component {
     render() {
         return (
             <div>
-                <ContractsTable/>
+                <Card style={styles.card}>
+                    <CardHeader title="Pending Contracts"/>
+
+                    <CardText>
+                        <ContractsTable complete={false}/>
+                    </CardText>
+                </Card>
+
+                <Card style={styles.card}>
+                    <CardHeader title="Completed Contracts"/>
+
+                    <CardText>
+                        <ContractsTable complete={true}/>
+                    </CardText>
+                </Card>
             </div>
         );
     }